# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "REST API для управления списком задач",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список задач",
                "operationId": "7786da811325606208249ebc8d393d9a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка (due_date, create_date, created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "operationId": "a737e990cef5769592bc41ee6a62f7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Задача1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-20T15:00:00"
                                    },
                                    "create_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-20T15:00:00"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ]
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Работа"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить конкретную задачу",
                "operationId": "dd21301b6ef0425f2260a34eac9679c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные задачи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "operationId": "2a0a8c212f84263f9a249469557b2539",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача обновлена"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "a2d3ec140e69f8b97ce812eadbe932ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача удалена"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "description": "Модель задачи",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "title": {
                        "description": "Название задачи",
                        "type": "string",
                        "example": "Купить хлеб"
                    },
                    "description": {
                        "description": "Описание задачи",
                        "type": "string",
                        "example": "Зайти в булочную после работы"
                    },
                    "due_date": {
                        "description": "Срок выполнения",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-20T15:00:00Z"
                    },
                    "create_date": {
                        "description": "Дата создания (из запроса)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-20T15:00:00Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "не выполнена",
                        "enum": [
                            "выполнена",
                            "не выполнена"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "средний",
                        "enum": [
                            "низкий",
                            "средний",
                            "высокий"
                        ]
                    },
                    "category": {
                        "description": "Категория",
                        "type": "string",
                        "example": "Дом"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Операции с задачами"
        }
    ]
}